Add VariableGroup model
Also moved variable model to new file and fixed imports
