+++ doku/models/base.py
@@ -1,8 +1,18 @@
-from doku.models.document import Document, Variable
+from doku.models.user import User
+from doku.models.document import Document
+from doku.models.variable import Variable, VariableGroup
 from doku.models.resource import Resource
 from doku.models.snippet import Snippet
 from doku.models.template import Template, Stylesheet
-from doku.models.user import User
 
 
-__all__ = [User, Document, Variable, Template, Stylesheet, Resource, Snippet]
+__all__ = [
+    User,
+    Document,
+    Variable,
+    VariableGroup,
+    Template,
+    Stylesheet,
+    Resource,
+    Snippet
+]
