+++ doku/models/schemas/template.py
@@ -7,7 +7,7 @@ from marshmallow_sqlalchemy.fields import Nested
 from werkzeug.exceptions import BadRequest
 
 from doku.models import DateSchemaMixin, db
-from doku.models.document import Variable
+from doku.models.variable import Variable
 from doku.models.schemas.common import ApiSchemaMixin, DokuSchema, NotEmptyString
 from doku.models.template import Template, Stylesheet
 from doku.utils.db import get_or_create
